requirements.txt
python-telegram-bot==21.6
apscheduler==3.10.4
python-dotenv==1.0.1
psycopg[binary]
flask==3.0.0